PM-18068 Addressed PR comments updating images to preserve vector data, some comment fixes, refactor of cipher state collections naming and computed properties.
