+++ BitwardenShared/UI/Platform/Application/Support/Images.xcassets/Icons/Cards/mastercard.imageset/Contents.json
@@ -18,5 +18,8 @@
   "info" : {
     "author" : "xcode",
     "version" : 1
+  },
+  "properties" : {
+    "preserves-vector-representation" : true
   }
 }
