+++ BitwardenShared/UI/Platform/Application/Support/Images.xcassets/Icons/folder16.imageset/Contents.json
@@ -10,6 +10,7 @@
     "version" : 1
   },
   "properties" : {
+    "preserves-vector-representation" : true,
     "template-rendering-intent" : "template"
   }
 }
